Start splitting out the different modes
